+++ packages/moonstone/MoonstoneDecorator/MoonstoneDecorator.js
@@ -5,10 +5,10 @@
  */
 
 import {addAll} from '@enact/core/keymap';
-import classnames from 'classnames';
 import hoc from '@enact/core/hoc';
 import I18nDecorator from '@enact/i18n/I18nDecorator';
 import React from 'react';
+import classNames from 'classnames';
 import {ResolutionDecorator} from '@enact/ui/resolution';
 import {FloatingLayerDecorator} from '@enact/ui/FloatingLayer';
 import SpotlightRootDecorator from '@enact/spotlight/SpotlightRootDecorator';
@@ -28,16 +28,17 @@ import {configure} from '@enact/ui/Touchable';
  * @hocconfig
  */
 const defaultConfig = {
-	i18n: true,
+	disableFullscreen: false,
 	float: true,
+	i18n: true,
 	noAutoFocus: false,
 	overlay: false,
 	ri: {
 		screenTypes
 	},
+	skin: true,
 	spotlight: true,
-	textSize: true,
-	skin: true
+	textSize: true
 };
 
 /**
@@ -59,10 +60,14 @@ const defaultConfig = {
  * @public
  */
 const MoonstoneDecorator = hoc(defaultConfig, (config, Wrapped) => {
-	const {ri, i18n, spotlight, float, noAutoFocus, overlay, textSize, skin, highContrast} = config;
+	const {ri, i18n, spotlight, float, noAutoFocus, overlay,
+		textSize, skin, highContrast, disableFullscreen} = config;
 
 	// Apply classes depending on screen type (overlay / fullscreen)
-	const bgClassName = 'enact-fit' + (overlay ? '' : ` ${css.bg}`);
+	const bgClassName = classNames({
+		'enact-fit': !disableFullscreen,
+		[css.bg]: !overlay
+	});
 
 	let App = Wrapped;
 	if (float) App = FloatingLayerDecorator({wrappedClassName: bgClassName}, App);
@@ -119,18 +124,13 @@ const MoonstoneDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		static displayName = 'MoonstoneDecorator';
 
 		render () {
-			const {className, ...rest} = this.props,
-				classes = [css.root, 'enact-unselectable', 'enact-fit'];
-
-			if (!float) {
-				classes.push(bgClassName);
-			}
-			if (className) {
-				classes.push(className);
-			}
+			const className = classNames(css.root, this.props.className, 'enact-unselectable', {
+				[bgClassName]: !float,
+				'enact-fit': !disableFullscreen
+			});
 
 			return (
-				<App {...rest} className={classnames(classes)} />
+				<App {...this.props} className={className} />
 			);
 		}
 	};