+++ packages/sampler/stories/qa-stories/Header.js
@@ -3,14 +3,17 @@ import Button from '@enact/moonstone/Button';
 import Input from '@enact/moonstone/Input';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {boolean, text} from '@storybook/addon-knobs';
+
+import {boolean, text} from '../../src/enact-knobs';
+
+Header.displayName = 'Header';
 
 storiesOf('Header', module)
 	.add(
 		'with headerComponent',
 		() => (
 			<Header
-				title={text('title', 'Title')}
+				title={text('title', Header, 'Title')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -21,7 +24,7 @@ storiesOf('Header', module)
 		() => (
 			<Header
 				type="compact"
-				title={text('title', 'Title')}
+				title={text('title', Header, 'Title')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -31,9 +34,9 @@ storiesOf('Header', module)
 		'with titles below',
 		() => (
 			<Header
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			/>
 		)
 	)
@@ -42,9 +45,9 @@ storiesOf('Header', module)
 		() => (
 			<Header
 				type="compact"
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			/>
 		)
 	)
@@ -52,9 +55,9 @@ storiesOf('Header', module)
 		'with titles below and headerComponent',
 		() => (
 			<Header
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -65,9 +68,9 @@ storiesOf('Header', module)
 		() => (
 			<Header
 				type="compact"
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -77,9 +80,9 @@ storiesOf('Header', module)
 		'with RTL text',
 		() => (
 			<Header
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'כתוביות למט')}
-				subTitleBelow={text('subTitleBelow', 'כתוביות למט')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'כתוביות למט')}
+				subTitleBelow={text('subTitleBelow', Header, 'כתוביות למט')}
 			/>
 		)
 	)
@@ -88,8 +91,8 @@ storiesOf('Header', module)
 		() => (
 			<Header
 				type="compact"
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'כתוביות למט')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'כתוביות למט')}
 			/>
 		)
 	)
@@ -97,9 +100,9 @@ storiesOf('Header', module)
 		'with long text and headerComponent',
 		() => (
 			<Header
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -110,9 +113,9 @@ storiesOf('Header', module)
 		() => (
 			<Header
 				type="compact"
-				title={text('title', 'Title')}
-				titleBelow={text('titleBelow', 'This is a header sample with long titleBelow text and header components to test positioning of header components.')}
-				subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+				title={text('title', Header, 'Title')}
+				titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow text and header components to test positioning of header components.')}
+				subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 			>
 				<Button small>On / Off</Button>
 			</Header>
@@ -122,13 +125,13 @@ storiesOf('Header', module)
 	.add(
 		'with Input, long text and headerComponent',
 		() => {
-			const input = boolean('Input Mode', true) ? <Input dismissOnEnter={boolean('Input dismissOnEnter', true)} /> : null;
+			const input = boolean('Input Mode', Header, true) ? <Input dismissOnEnter={boolean('Input dismissOnEnter', Header, true)} /> : null;
 			return (
 				<Header
-					title={text('title', 'Title')}
+					title={text('title', Header, 'Title')}
 					headerInput={input}
-					titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-					subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+					titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+					subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 				>
 					<Button small>On / Off</Button>
 				</Header>
@@ -138,13 +141,13 @@ storiesOf('Header', module)
 	.add(
 		'with Input, tall-glyphs, and titles below',
 		() => {
-			const input = boolean('Input Mode', true) ? <Input dismissOnEnter={boolean('Input dismissOnEnter', true)} /> : null;
+			const input = boolean('Input Mode', Header, true) ? <Input dismissOnEnter={boolean('Input dismissOnEnter', Header, true)} /> : null;
 			return (
 				<Header
-					title={text('title', 'ิ้  ไั  ஒ  து': 'ิ้  ไั  ஒ  த')}
+					title={text('title', Header, 'ิ้  ไั  ஒ  து': 'ิ้  ไั  ஒ  த')}
 					headerInput={input}
-					titleBelow={text('titleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
-					subTitleBelow={text('subTitleBelow', 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+					titleBelow={text('titleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
+					subTitleBelow={text('subTitleBelow', Header, 'This is a header sample with long titleBelow, subTitleBelow text and header components to test positioning of header components.')}
 				/>
 			);
 		}