+++ packages/sampler/stories/qa-stories/SwitchItem.js
@@ -4,7 +4,10 @@ import Divider from '@enact/moonstone/Divider';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {text, boolean} from '@storybook/addon-knobs';
+
+import {boolean, text} from '../../src/enact-knobs';
+
+SwitchItem.displayName = 'SwitchItem';
 
 const inputData = {
 	longText : 'Looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong Text',
@@ -18,18 +21,18 @@ storiesOf('SwitchItem', module)
 		() => (
 			<div>
 				<SwitchItem
-					disabled={boolean('disabled', false)}
-					inline={boolean('inline', false)}
+					disabled={boolean('disabled', SwitchItem, false)}
+					inline={boolean('inline', SwitchItem, false)}
 					onToggle={action('onToggle')}
 				>
-					{text('Long Text', inputData.longText)}
+					{text('Long Text', SwitchItem, inputData.longText)}
 				</SwitchItem>
 				<SwitchItem
 					disabled
-					inline={boolean('inline', false)}
+					inline={boolean('inline', SwitchItem, false)}
 					onToggle={action('onToggle')}
 				>
-					{text('Disable Long Text', inputData.disabledLong)}
+					{text('Disable Long Text', SwitchItem, inputData.disabledLong)}
 				</SwitchItem>
 			</div>
 		)
@@ -44,14 +47,14 @@ storiesOf('SwitchItem', module)
 				<Group
 					childComponent={SwitchItem}
 					itemProps={{
-						inline: boolean('ItemProps-Inline', false),
-						disabled: boolean('disabled', false)
+						inline: boolean('ItemProps-Inline', SwitchItem, false),
+						disabled: boolean('disabled', SwitchItem, false)
 					}}
 					selectedProp="selected"
 					defaultSelected={1}
 					onSelect={action('onSelect')}
 				>
-					{[text('Normal Text 1', inputData.normalText + 1), text('Normal Text 2', inputData.normalText + 2), text('Normal Text 3', inputData.normalText + 3)]}
+					{[text('Normal Text 1', SwitchItem, inputData.normalText + 1), text('Normal Text 2', SwitchItem, inputData.normalText + 2), text('Normal Text 3', SwitchItem, inputData.normalText + 3)]}
 				</Group>
 				<Divider>
 					{'Switch items with long text in a group'}
@@ -59,16 +62,16 @@ storiesOf('SwitchItem', module)
 				<Group
 					childComponent={SwitchItem}
 					itemProps={{
-						inline: boolean('ItemProps-Inline', false),
-						disabled: boolean('disabled', false)
+						inline: boolean('ItemProps-Inline', SwitchItem, false),
+						disabled: boolean('disabled', SwitchItem, false)
 					}}
 					childSelect="onToggle"
 					selectedProp="selected"
-					disabled={boolean('disabled', false)}
+					disabled={boolean('disabled', SwitchItem, false)}
 					defaultSelected={1}
 					onSelect={action('onSelect')}
 				>
-					{[text('Long Text 1', 'First ' + inputData.longText), text('Long Text 2', 'Second ' + inputData.longText), text('Long Text 3', 'Third ' + inputData.longText)]}
+					{[text('Long Text 1', SwitchItem, 'First ' + inputData.longText), text('Long Text 2', SwitchItem, 'Second ' + inputData.longText), text('Long Text 3', SwitchItem, 'Third ' + inputData.longText)]}
 				</Group>
 			</div>
 		)